convertor/backend/pyproject.toml
[project]
name = "markdown-docs-converter"
version = "1.0.0"
description = "SOTA Markdown to Documentation Website Converter"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "markdown-it-py>=3.0.0",
    "mdit-py-plugins>=0.4.0",
    "pygments>=2.17.0",
    "pyyaml>=6.0.1",
    "watchfiles>=0.21.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "docutils>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
